+++ packages/core/src/swarm/index.ts
@@ -1,6 +1,8 @@
 export * from "./complete-cohort.js";
 export { completeCohortMain } from "./complete-cohort-cli.js";
 export * from "./constants.js";
+export * from "./finalize-cohort.js";
+export { finalizeCohortMain, parseFinalizeCohortArgv } from "./finalize-cohort-cli.js";
 export * from "./launch.js";
 export { launchMain, parseLaunchArgv } from "./launch-cli.js";
 export * from "./readiness.js";